Init package structure
