+++ nirdizati/encoders/__init__.py
@@ -1,0 +1,5 @@
+from nirdizati.encoders.boolean_encoder import BooleanEncoder
+from nirdizati.encoders.complex_encoder import ComplexEncoder
+from nirdizati.encoders.frequency_encoder import FrequencyEncoder
+from nirdizati.encoders.index_latest_payload_encoder import IndexLatestPayloadEncoder
+from nirdizati.encoders.simple_index_encoder import SimpleIndexEncoder
