+++ nirdizati/encoders/boolean_encoder.py
@@ -1,12 +1,8 @@
-import csv
-import datetime
-import json
-import time
-import encoder
-from os.path import isfile
-
-import pandas as pd
 import numpy as np
+import pandas as pd
+
+import encoder
+
 
 class BooleanEncoder:
