+++ setup.py
@@ -1,0 +1,9 @@
+from setuptools import setup
+
+setup(name='nirdizati.encoders',
+      version='0.1',
+      description='Encoders for Nirdizati training',
+      url='https://github.com/nirdizati/nirdizati-encoders',
+      author='TKasekamp',
+      packages=['nirdizati.encoders'],
+      zip_safe=False)
